booke: timer: clear only TCR on guest reset
The timer reset handler is calling store_booke_tcr() to
reset TCR to 0. The store_booke_tcr() clears the TCR but also
calls booke_update_fixed_timer(), which modify/start the qemu
timer to fire when new time calculated with new TCR (which is 0).

But on reset we do not want to start/modify the qemu timer.

Signed-off-by: Bharat Bhushan <bharat.bhushan@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>